# api-service/internal/items/http/app/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_abdulazizax_mini-twitter_api-service_genproto_like.GetLikesResponse:
    properties:
      likes:
        description: List of  likes
        items:
          $ref: '#/definitions/github_com_abdulazizax_mini-twitter_api-service_genproto_like.LikeInfo'
        type: array
    type: object
  github_com_abdulazizax_mini-twitter_api-service_genproto_like.LikeInfo:
    properties:
      liked_at:
        allOf:
        - $ref: '#/definitions/timestamppb.Timestamp'
        description: Timestamp when the like was added
      username:
        description: username of the user who liked
        type: string
    type: object
...
  timestamppb.Timestamp:
    properties:
      nanos:
        description: |-
          Non-negative fractions of a second at nanosecond resolution. Negative
          second values with fractions must still have non-negative nanos values
          that count forward in time. Must be from 0 to 999,999,999
          inclusive.
        type: integer
      seconds:
        description: |-
          Represents seconds of UTC time since Unix epoch
          1970-01-01T00:00:00Z. Must be from 0001-01-01T00:00:00Z to
          9999-12-31T23:59:59Z inclusive.
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API Endpoints for MiniTwitter
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: '# MiniTwitter'
  version: 1.03.67.83.145
paths:
...
  /user/tweet/comment/like/{comment_id}:
    get:
      consumes:
      - application/json
      description: Retrieves the likes for a specific comment
      parameters:
      - description: Comment ID
        in: path
        name: comment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_abdulazizax_mini-twitter_api-service_genproto_like.GetLikesResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get likes for a comment
      tags:
      - Likes
    post:
      consumes:
      - application/json
      description: Adds a like to a comment for the authenticated user
      parameters:
      - description: Comment ID
        in: path
        name: comment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Like a comment
      tags:
      - Likes
  /user/tweet/comment/like/{comment_id}/unlike:
    post:
      consumes:
      - application/json
      description: Removes a like from a comment for the authenticated user
      parameters:
      - description: Comment ID
        in: path
        name: comment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Unlike a comment
      tags:
      - Likes
  /user/tweet/like/{tweet_id}:
    get:
      consumes:
      - application/json
      description: Retrieves the likes for a specific tweet
      parameters:
      - description: Tweet ID
        in: path
        name: tweet_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_abdulazizax_mini-twitter_api-service_genproto_like.GetLikesResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get likes for a tweet
      tags:
      - Likes
    post:
      consumes:
      - application/json
      description: Adds a like to a tweet for the authenticated user
      parameters:
      - description: Tweet ID
        in: path
        name: tweet_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Like a tweet
      tags:
      - Likes
  /user/tweet/like/{tweet_id}/unlike:
    post:
      consumes:
      - application/json
      description: Removes a like from a tweet for the authenticated user
      parameters:
      - description: Tweet ID
        in: path
        name: tweet_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Unlike a tweet
      tags:
      - Likes
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"